+++ CHANGES.rst
@@ -5,6 +5,11 @@ Changelog
 Unreleased
 ----------
 
+* Temporarily set ``cluster.routing.allocation.enable`` to ``new_primaries`` before
+  deleting each pod and reset it after the pod has restarted. This ensures more
+  controlled shard allocation during rolling restarts.
+
+
 2.47.1 (2025-05-14)
 -------------------